+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fix dropdown menu buttons margins
 * Fix dashboard now values now only use capacity and realtime, not bookings
 * Settings page now restricted to administrators 
+* Fix timeslots, remove the first option "Pick a timeslot" to let appear first "No timeslot available"
 
 ### 1.4.3: May 24th, 2021
 * Times dropdown must be above header in fixed position